Refactored and cleaned up (GH-45647)
