+++ pandas/tests/window/conftest.py
@@ -2,12 +2,10 @@ from datetime import (
     datetime,
     timedelta,
 )
-from functools import partial
 
 import numpy as np
 import pytest
 
-import pandas._libs.window.aggregations as window_aggregations
 import pandas.util._test_decorators as td
 
 from pandas import (
@@ -261,62 +259,3 @@ def frame():
         index=bdate_range(datetime(2009, 1, 1), periods=100),
         columns=np.arange(10),
     )
-
-
-def _named_func(name_and_func):
-    name, func = name_and_func
-    if not hasattr(func, "func"):
-        func = partial(func)
-    func.__name__ = name
-    return func
-
-
-@pytest.fixture(
-    params=[
-        _named_func(x)
-        for x in [
-            ("roll_sum", window_aggregations.roll_sum),
-            ("roll_mean", window_aggregations.roll_mean),
-        ]
-        + [
-            (f"roll_var({ddof})", partial(window_aggregations.roll_var, ddof=ddof))
-            for ddof in [0, 1]
-        ]
-        + [
-            ("roll_skew", window_aggregations.roll_skew),
-            ("roll_kurt", window_aggregations.roll_kurt),
-            ("roll_median_c", window_aggregations.roll_median_c),
-            ("roll_max", window_aggregations.roll_max),
-            ("roll_min", window_aggregations.roll_min),
-        ]
-        + [
-            (
-                f"roll_quantile({quantile},{interpolation})",
-                partial(
-                    window_aggregations.roll_quantile,
-                    quantile=quantile,
-                    interpolation=interpolation,
-                ),
-            )
-            for quantile in [0.0001, 0.5, 0.9999]
-            for interpolation in window_aggregations.interpolation_types
-        ]
-        + [
-            (
-                f"roll_rank({percentile},{method},{ascending})",
-                partial(
-                    window_aggregations.roll_rank,
-                    percentile=percentile,
-                    method=method,
-                    ascending=ascending,
-                ),
-            )
-            for percentile in [True, False]
-            for method in window_aggregations.rolling_rank_tiebreakers.keys()
-            for ascending in [True, False]
-        ]
-    ]
-)
-def rolling_aggregation(request):
-    """Make a named rolling aggregation function as fixture."""
-    return request.param
