+++ pandas/tests/window/test_rolling.py
@@ -2,7 +2,6 @@ from datetime import (
     datetime,
     timedelta,
 )
-import sys
 
 import numpy as np
 import pytest
@@ -1735,38 +1734,3 @@ def test_rolling_std_neg_sqrt():
 
     b = a.ewm(span=3).std()
     assert np.isfinite(b[2:]).all()
-
-
-def test_rolling_aggregation_boundary_consistency(rolling_aggregation):
-    # GH-45647
-    minp, step, width, size, selection = 0, 1, 3, 11, [2, 7]
-    s = Series(np.arange(1, 1 + size, dtype=np.float64))
-    end = np.arange(width, size, step, dtype=np.int64)
-    start = end - width
-    selarr = np.array(selection, dtype=np.int32)
-    result = Series(rolling_aggregation(s.values, start[selarr], end[selarr], minp))
-    expected = Series(rolling_aggregation(s.values, start, end, minp)[selarr])
-    tm.assert_equal(expected, result)
-
-
-def test_rolling_aggregation_with_unused_elements(rolling_aggregation):
-    # GH-45647
-    minp, width = 0, 5  # width at least 4 for kurt
-    size = 2 * width + 5
-    s = Series(np.arange(1, size + 1, dtype=np.float64))
-    s[width : width + 2] = sys.float_info.min
-    s[width + 2] = np.nan
-    s[width + 3 : width + 5] = sys.float_info.max
-    start = np.array([0, size - width], dtype=np.int64)
-    end = np.array([width, size], dtype=np.int64)
-    loc = np.array(
-        [j for i in range(len(start)) for j in range(start[i], end[i])],
-        dtype=np.int32,
-    )
-    result = Series(rolling_aggregation(s.values, start, end, minp))
-    compact_s = np.array(s.iloc[loc], dtype=np.float64)
-    compact_start = np.arange(0, len(start) * width, width, dtype=np.int64)
-    compact_end = compact_start + width
-    expected = Series(rolling_aggregation(compact_s, compact_start, compact_end, minp))
-    assert np.isfinite(expected.values).all(), "Not all expected values are finite"
-    tm.assert_equal(expected, result)
